Only update job status to failed from Twilio
Updating every message status from Twilio is not needed, and creating
a lot of unnecessary jobs.

We only care if it's failed, so we only update the job status
when the message status is failed.
